feat(sign-in): refactor sign-in component and add signup route
