--- src/app/features/auth/signin/sign-in.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SignIn } from './sign-in';
-
-describe('SignIn', () => {
-  let component: SignIn;
-  let fixture: ComponentFixture<SignIn>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      imports: [SignIn]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(SignIn);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
